+++ src/server/generic.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use std::collections::HashMap;
 
-use crate::config::{GlobalConfig, ServerConfig};
+use crate::config::{BackupConfig, GlobalConfig, ServerConfig};
 
 use super::manager::ServerManager;
 
@@ -14,9 +14,10 @@ impl GenericServer {
         server_id: String,
         config: ServerConfig,
         global: &GlobalConfig,
+        backup: &BackupConfig,
         java_environments: &HashMap<String, String>,
     ) -> Result<Self> {
-        let manager = ServerManager::new(server_id, config, global, java_environments)?;
+        let manager = ServerManager::new(server_id, config, global, backup, java_environments)?;
         Ok(Self { manager })
     }
 }